Move builder to subdirectory
