+++ MagicGradients/Builder/IChildBuilder.cs
@@ -1,0 +1,8 @@
+﻿namespace MagicGradients
+{
+    public interface IChildBuilder
+    {
+        StopsFactory StopsFactory { get; }
+        Gradient Construct();
+    }
+}
